Warning count to 1356 (Linux)

+++ src/MEGASync/gui/GuestWidget.cpp
@@ -269,8 +269,8 @@ void GuestWidget::onRequestUpdate(MegaApi*, MegaRequest *request)
     {
         if (request->getTotalBytes() > 0)
         {
-            ui->progressBar->setMaximum(request->getTotalBytes());
-            ui->progressBar->setValue(request->getTransferredBytes());
+            ui->progressBar->setMaximum(static_cast<int>(request->getTotalBytes()));
+            ui->progressBar->setValue(static_cast<int>(request->getTransferredBytes()));
         }
     }
 }
